Clean up some unused bash aliases

+++ .bash_aliases
@@ -42,40 +42,27 @@ alias scrot='scrot -c -d 7'
 alias show-info='~/.bin/info.pl'
 alias show-colors='~/.bin/colors.sh'
 
-## Sudo fixes
-alias orphand='sudo deborphan | xargs sudo apt-get -y remove --purge'
-alias cleanup='sudo apt-get autoclean && sudo apt-get autoremove && sudo apt-get clean && sudo apt-get remove && orphand'
-alias updatedb='sudo updatedb'
-
-## Dev related
-alias restart-apache='sudo /etc/init.d/apache2 restart'
-
 ## Misc
-alias edit='vim'
 alias pg='ps -ef | grep'
-alias colour='/bin/bash ~/workspace/personal/dot_files/256_colors.sh'
-
-## apt-get related
-alias si='sudo apt-get install'
-alias sd='sudo apt-get update'
-
-# ssh connections
-alias jss='ssh vincent@jinshisong.com'
-alias yb='ssh vinnie@yongbet.com'
-alias stg='ssh vincent@staging.jinshisong.com'
-alias mt='ssh vincent@211.155.86.126'
 
 # some development quick n dirty shortcuts
-alias s='cd /home/vincent/workspace/jss/spree'
-alias h='cd /home/vincent/workspace/jss/horsery'
-alias ccache='/bin/rm -r /home/vincent/workspace/jss/spree/tmp/* /home/vincent/workspace/jss/spree/public/assets/*'
-alias pryc='bundle exec pry -r ./config/environment'
+alias cap='bundle exec cap'
 alias rake='bundle exec rake'
-alias r='rails r'
-alias hs='python -m SimpleHTTPServer'
-alias dbreload='rake db:drop db:create db:migrate'
-alias sync_production='git co master; rake db:drop db:create; cap production db:sync_to_local'
-alias sync_stating='git co master; rake db:drop db:create; cap staging db:sync_to_local'
+alias pyserver='python -m SimpleHTTPServer'
 
-alias vi='$(which vim)'
-alias vim='$(which vim)'
+alias vi='vim'
+
+# OSX specific aliases
+if [[ $OSTYPE == darwin* ]]; then
+  alias gvim='mvim'
+
+else
+  # Sudo fixes
+  alias orphand='sudo deborphan | xargs sudo apt-get -y remove --purge'
+  alias cleanup='sudo apt-get autoclean && sudo apt-get autoremove && sudo apt-get clean && sudo apt-get remove && orphand'
+  alias updatedb='sudo updatedb'
+
+  # apt-get
+  alias si='sudo apt-get install'
+  alias sd='sudo apt-get update'
+fi
